[tcat] add TCAT_ENABLE.req TMF command
This adds support for the TMF command to enable TCAT remotely.
A test is added that uses the 'UDP send' mechanism to send the new TMF command
to a target node.

Some fixes/additions to the test framework are made to support the new test,
including a new argument for udp_send() to send a specific byte array and
udp_rx() to receive data by a UDP client on a node.