[tcat][ble] improve TCAT Commissioner output logging for SSL errors; fix lib dependencies
This PR provides more structured logging for ssl.py errors, and displays the OpenSSL verify error code.
This is used for certification to validate reasons of handshake failure.

Also, it corrects a dependency problem: the 'cryptography' library was not imported from pyproject.toml.
This causes the app to fail to run on some platforms that didn't manually install this library.
Due to these major dependency changes, the version of the app is raised to 0.2.0.