[radio] add radio availability map updated API
When Thread is using the same radio chip with BT or 2.4GHz WiFi,
the radio driver has the ability to know when BT or WiFi will
occupy the radio. To reduce the interference between Thread and
BT or WiFi, the radio driver calls this method to notify OpenThread
when the radio is available for Thread. OpenThread will try its
best to avoid using the not allowed radio slots.

This commit adds a radio platform API `otPlatRadioAvailMapUpdated()`
for the radio driver to report the radio availability map to OpenThread.