Expose multi-instance getter as instance API
This commit adds a public getter function to acquire a pointer for
an OpenThread instance associated with a provided index.

An index to instance reference getter function was previously implemented
in the core instance interface, but this capability was not publicly accessible.
The accessor function introduced in this commit will now allow for platforms
or application code to perform this mapping.