[tcat] enable TCAT Commissioner to receive Alerts/TLV events over TLS and improve connection mgmt
This enables the TCAT Commissioner to receive data such as TLS Alerts, or asynchronously sent 'event' TLVs, over TLS.
Processing TLS Alert is required to detect the sending of Alert by the TCAT Device, which is a requirement to be
verified in cert tests. An async background process is started to receive and log the received events.

Also some minor improvements in connection state management: when certain commands are given after the TCAT link
is disconnected, or when a TCAT link could not be established, a message will be printed to clearly say it's
disconnected, instead of a cryptic error.

The CA certificate store for CommCert3 is extended with an additional CA certificate, so that it can be verified
in cert tests that a TCAT Device rejects a wrong Commissioner with a TLS Alert (previously this couldn't be tested).

Also includes a fix of the pyproject.toml such that Poetry does not display the long warning on installation.

Also includes an improvement of TLV displaying to the user with a STRING field, if the value is a string.