[udp] no changing netif of bound sockets
The UDP socket may or may not be created from the platform layer based
on the network interface specified when Udp::Open(), changing the
network would result in inconsistent states. This commit defers creating
platform socket in Udp::Bind() and disallows changing the network
interface of a bound socket to make sure the states are consistent in
OpenThread stack and the platform stack.