[ncp] add netif state property change (#11950)
This commit makes NCP send notification to the host when the netif
state changes.

At the beginning we use ThreadHost API to start the thread network on
OTBR NCP which explicitly sends spinel property set command to the NCP
and get response. So the host side can get the latest state of netif
in the response. However if we use the cli command (`ifconfig up`) to
start the thread network, we need the NCP to send a notification to
the host that netif state changes.

I have verified that with the change, the host can receive the update
if we issue `ifconfig up` command.