Expose instance getter functions in instance API
This commit adds public getter functions to acquire a pointer for
the single OpenThread instance or one associated with a provided index.

An index to instance reference getter function was previously implemented
in the core instance interface, but this capability was not publicly accessible.
The accessor function introduced in this commit will now allow for platforms
or application code to perform this mapping. Corresponding changes for the
single instance case are also included.