[trel] simplify `Peer::Matches()` for name matching
Consolidates the two overloaded `Peer::Matches()` methods for service
and host name matching into a single method.

This change introduces a new `NameMatchType` enum, which is used to
specify whether to match against a service name or a host name. This
removes the need for the now-unnecessary `ServiceNameMatcher` and
`HostNameMatcher` structs, simplifying the calling code and the `Peer`
class.