[tcat] fix missing Python modules in tcat-ble-client pyproject.toml; bump module versions
This fixes the issue that some required Python modules were missing in pyproject.toml. To avoid Python version conflicts with
the cryptography module, the minimum Python version is raised to 3.10. Also, the name, version and description of the project
are updated. Module versions are updated to recent (2025) versions.

For Windows, a platform-specific import of pyreadline3 is added. Due to this conditional dependency, the poetry.lock file is
removed from the repository: it is now necessarily specific to each platform and has to be locally generated.