[ble][tcat] improve error handling for BLE packets and TCAT TLVs
Updates in error handling to ensure that recoverable BLE errors are logged, failures don't lead to partial
TLVs being generated or processed, and internal allocation failures lead to a GeneralError status response
back to the TCAT Commissioner. In case of likely unrecoverable TLV damage, or cases where the GeneralError
status response even couldn't be generated, the secure TLS connection is closed (Disconnect).

BLE ATT MTU handling is updated to use default MTU 23 if fetching the MTU fails; and more clear check on
min MTU. If the platform reports a very high MTU (above max) this is not seen as an error, but BleSecure
will simply uses the highest usable MTU that it has configured, whether the platform's report is correct
or not.

Close #11536