[mle] delay child update tx to restore rx-on children
This commit introduces a mechanism to schedule and retry the
transmission of "Child Update Request" messages to restore
rx-on-when-idle (non-sleepy) children upon role restoration.
It spreads the transmissions using a gap interval with jitter
(15 ± 2 msec) to avoid traffic bursts. It also includes a
retry mechanism starting with 2 seconds delay with exponential
backoff up to a maximum interval of 64 seconds if responses
are not received.

The new `Mle::ScheduleChildUpdateToRestoreNonSleepyChildren` method
is called when a router or leader role is assumed, and retries are
managed within the `HandleTimeTick`. Sleepy children continue to be
handled and recovered as before during time ticks.